+++ boot/src/test/kotlin/io/github/costaalex/workoutrelay/infrastructure/platform/trainerroad/workout/TrainerRoadWorkoutRepositoryTest.kt
@@ -34,7 +34,7 @@ class TrainerRoadWorkoutRepositoryTest {
         TrainerRoadApiClientService(
             trainerRoadApiClient,
             trainerRoadConfigurationRepository,
-            TrainerRoadWorkoutCacheService(trainerRoadApiClient, trainerRoadConfigurationRepository, objectMapper),
+            TrainerRoadWorkoutCacheService(trainerRoadApiClient, trainerRoadConfigurationRepository),
         )
 
     private val trainerRoadWorkoutRepository =